Get file tree data in dashboard and save in redux store
